requirements.txt
tavily-python
google-generativeai
groq
python-dotenv
colorama